+++ extensions/shared/library/src/main/java/app/revanced/extension/shared/Utils.java
@@ -40,13 +40,15 @@ import java.util.concurrent.SynchronousQueue;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
+import app.revanced.extension.shared.settings.AppLanguage;
+import app.revanced.extension.shared.settings.BaseSettings;
 import app.revanced.extension.shared.settings.BooleanSetting;
 import app.revanced.extension.shared.settings.preference.ReVancedAboutPreference;
 
 public class Utils {
 
     @SuppressLint("StaticFieldLeak")
-    private static Context context;
+    private static volatile Context context;
 
     private static String versionName;
     private static String applicationLabel;
@@ -360,7 +362,17 @@ public class Utils {
     }
 
     public static void setContext(Context appContext) {
+        // Must initially set context as the language settings needs it.
         context = appContext;
+
+        AppLanguage language = BaseSettings.REVANCED_LANGUAGE.get();
+        if (language != AppLanguage.DEFAULT) {
+            // Create a new context with the desired language.
+            Configuration config = appContext.getResources().getConfiguration();
+            config.setLocale(language.getLocale());
+            context = appContext.createConfigurationContext(config);
+        }
+
         // In some apps like TikTok, the Setting classes can load in weird orders due to cyclic class dependencies.
         // Calling the regular printDebug method here can cause a Settings context null pointer exception,
         // even though the context is already set before the call.
@@ -523,6 +535,11 @@ public class Utils {
         return currentNightMode == Configuration.UI_MODE_NIGHT_YES;
     }
 
+    public static boolean isLandscapeOrientation() {
+        final int orientation = context.getResources().getConfiguration().orientation;
+        return orientation == Configuration.ORIENTATION_LANDSCAPE;
+    }
+
     /**
      * Automatically logs any exceptions the runnable throws.
      *
@@ -595,7 +612,7 @@ public class Utils {
                 || networkType == NetworkType.OTHER;
     }
 
-    @SuppressLint("MissingPermission") // permission already included in YouTube
+    @SuppressLint({"MissingPermission", "deprecation"}) // Permission already included in YouTube.
     public static NetworkType getNetworkType() {
         Context networkContext = getContext();
         if (networkContext == null) {
@@ -760,8 +777,8 @@ public class Utils {
             return;
         }
 
-        String deviceLanguage = Utils.getContext().getResources().getConfiguration().locale.getLanguage();
-        if (deviceLanguage.equals("en")) {
+        String revancedLocale = Utils.getContext().getResources().getConfiguration().locale.getLanguage();
+        if (revancedLocale.equals(Locale.ENGLISH.getLanguage())) {
             return;
         }
 
@@ -769,8 +786,8 @@ public class Utils {
             Preference pref = group.getPreference(i);
             pref.setSingleLineTitle(false);
 
-            if (pref instanceof PreferenceGroup) {
-                setPreferenceTitlesToMultiLineIfNeeded((PreferenceGroup) pref);
+            if (pref instanceof PreferenceGroup subGroup) {
+                setPreferenceTitlesToMultiLineIfNeeded(subGroup);
             }
         }
     }