+++ extensions/shared/library/src/main/java/app/revanced/extension/shared/spoof/requests/PlayerRoutes.java
@@ -10,7 +10,7 @@ import app.revanced.extension.shared.Logger;
 import app.revanced.extension.shared.requests.Requester;
 import app.revanced.extension.shared.requests.Route;
 import app.revanced.extension.shared.settings.BaseSettings;
-import app.revanced.extension.shared.spoof.AudioStreamLanguage;
+import app.revanced.extension.shared.settings.AppLanguage;
 import app.revanced.extension.shared.spoof.ClientType;
 
 final class PlayerRoutes {
@@ -42,9 +42,9 @@ final class PlayerRoutes {
             // but if this is a fall over client it will set the language even though
             // the audio language is not selectable in the UI.
             ClientType userSelectedClient = BaseSettings.SPOOF_VIDEO_STREAMS_CLIENT_TYPE.get();
-            AudioStreamLanguage language = userSelectedClient == ClientType.ANDROID_VR_NO_AUTH
+            AppLanguage language = userSelectedClient == ClientType.ANDROID_VR_NO_AUTH
                     ? BaseSettings.SPOOF_VIDEO_STREAMS_LANGUAGE.get()
-                    : AudioStreamLanguage.DEFAULT;
+                    : AppLanguage.DEFAULT;
 
             JSONObject client = new JSONObject();
             client.put("hl", language.getLanguage());