+++ extensions/youtube/stub/src/main/java/com/google/protos/youtube/api/innertube/InnertubeContext$ClientInfo.java
@@ -1,5 +1,0 @@
-package com.google.protos.youtube.api.innertube;
-
-public class InnertubeContext$ClientInfo {
-    public int r;
-}