+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/hide/shorts/HideShortsComponentsPatch.kt
@@ -71,6 +71,7 @@ private val hideShortsComponentsResourcePatch = resourcePatch {
             SwitchPreference("revanced_hide_shorts_home"),
             SwitchPreference("revanced_hide_shorts_subscriptions"),
             SwitchPreference("revanced_hide_shorts_search"),
+            SwitchPreference("revanced_hide_shorts_history"),
 
             PreferenceScreenPreference(
                 key = "revanced_shorts_player_screen",